requirements.txt
# requirements.txt
# Binance Futures Testnet Trading Bot dependencies

# HTTP client (used by BinanceClient for REST API calls)
requests==2.32.3

# Retry logic for HTTP adapters
urllib3==2.2.3

# Environment variable loading from .env files
python-dotenv==1.0.1
